+++ src/Symfony/Bundle/SecurityBundle/Tests/DependencyInjection/Fixtures/php/merge.php
@@ -3,7 +3,6 @@
 $this->load('merge_import.php');
 
 $container->loadFromExtension('security', [
-    'enable_authenticator_manager' => true,
     'providers' => [
         'default' => ['id' => 'foo'],
     ],